requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
blis==0.7.5
bs4==0.0.1
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cymem==2.0.6
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.47.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
idna==3.3
Jinja2==3.0.3
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.22.1
oauthlib==3.1.1
packaging==21.3
pathy==0.6.1
Pillow==9.0.0
preshed==3.0.6
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.7
PyQRCode==1.2.1
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wasabi==0.9.0
